[ui-size + feed-probe] user UI-size pref + nameplate distance 1.5x + chat-resurrect probe
User asks (2026-07-04): (1) "все менюшки и тексты ПОБОЛЬШЕ" -- ui.scale
user pref (default 1.25x, clamp 0.75-1.75) multiplied into the overlay
scale factor (combined cap 4.0); live slider in F1 > Cosmetics >
Interface, applies on release (atlas re-bake), persisted in
votv-coop.ini. (2) nameplate nicks unreadable too soon -- DistanceScale
kRefCm 600->900: full size to ~9 m, legibility floor at ~22.5 m (1.5x).
(3) the long-standing "expired chat line reappears for ~0.5 s" glitch --
static RE exhausted (store math cannot resurrect: per-entry alpha
monotone after the arrival ramp, expiry pops oldest, QueueSend hops to
the game thread, the host relay excludes the origin slot), so per the
probe-dont-guess rule chat_feed gains a PERMANENT probe: every push /
expiry / overflow drop logged, re-push of a text that died <60 s ago =
"feed: RESURRECT" warn, and an entry-keyed (bornMs) + fade-out-tail-
gated ALPHA-JUMP detector for the cant-happen case. Audit: 0 CRITICAL;
its two detector WARNs (fade-in false positive, same-text spam) closed
in this build; 2 pre-existing C4244 in event_feed wisp lambdas
cast-fixed.

DLL 68DE991B1FABCB74 deployed 4/4 hash-verified; wire v96 unchanged.
Verdict = runbook 0l.

+++ src/votv-coop/src/coop/session/event_feed.cpp
@@ -219,7 +219,7 @@ void Update(net::Session& session, void* localPlayer) {
             }
             net::WispGrabPayload p{};
             std::memcpy(&p, msg.payload, sizeof(p));
-            const uint8_t sender = msg.senderPeerSlot;
+            const uint8_t sender = static_cast<uint8_t>(msg.senderPeerSlot);
             ue_wrap::game_thread::Post([p, sender] { coop::wisp_tear_mirror::OnWispGrab(p, sender); });
             break;
         }
@@ -239,7 +239,7 @@ void Update(net::Session& session, void* localPlayer) {
             }
             net::WispTearPayload p{};
             std::memcpy(&p, msg.payload, sizeof(p));
-            const uint8_t sender = msg.senderPeerSlot;
+            const uint8_t sender = static_cast<uint8_t>(msg.senderPeerSlot);
             ue_wrap::game_thread::Post([p, sender] { coop::wisp_tear_mirror::OnWispTear(p, sender); });
             break;
         }
